+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ HEAD
 **To be included in next release**
 
   - add `perun fuzz` command which does a performance fuzzing
+  - remodel runner functions to generators
 
 0.16 (2019-02-16)
 -----------------